requirements.txt
cycler==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
pandas==1.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-mnist==0.7
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
